requirements.txt
Flask
nltk
numpy
tflearn
tensorflow
PyYAML
libretranslate
openai-whisper
